+++ changelog.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0)
   If you previously used the buttons `missing247` and `missing248`, please update to the new names.
 - Added more MacOS keys (#936)
 - Added keycodes above 255. If you are on linux you can use them now. (#935)
+- Added `boot.initrd.services.kmonad.enable` NixOS option to use KMonad in the initrd (#941).
 
 ### Changed